Create Graph data structure
Adjacency List:
An adjacency list is a simple way to represent a graph as a list of vertices, where each vertex has a list of its adjacent vertices. Here's an example of an adjacency list for an undirected graph with 4 vertices:
makefile
Copy code
0: 1 3
1: 0 2
2: 1 3
3: 0 2
In this example, vertex 0 is adjacent to vertices 1 and 3, vertex 1 is adjacent to vertices 0 and 2, and so on.

Adjacency Matrix:
An adjacency matrix is a two-dimensional array that represents the graph by storing a 1 at position (i,j) if there is an edge from vertex i to vertex j, and 0 otherwise. Here's an example of an adjacency matrix for the same undirected graph:
Copy code
  0 1 2 3
0 0 1 0 1
1 1 0 1 0
2 0 1 0 1
3 1 0 1 0
In this example, there is an edge from vertex 0 to vertex 1 (represented by a 1 in position (0,1)), an edge from vertex 1 to vertex 0 (represented by a 1 in position (1,0)), and so on.

Incidence Matrix:
An incidence matrix is a two-dimensional array that represents the graph by storing a 1 at position (i,j) if vertex i is incident on edge j, and 0 otherwise. Here's an example of an incidence matrix for the same undirected graph:
Copy code
  0 1 2 3
0 1 1 0 1
1 1 0 1 0
2 0 1 1 0
3 1 0 0 1
In this example, vertex 0 is incident on edges 0, 1, and 3 (represented by a 1 in positions (0,0), (0,1), and (0,3)), vertex 1 is incident on edges 0, 2 (represented by a 1 in positions (1,0) and (1,2)), and so on.

Each representation has its own advantages and disadvantages depending on the application, and choosing the right representation can have a significant impact on the performance of graph algorithms.